+++ nemo_skills/pipeline/app.py
@@ -22,7 +22,13 @@ import nemo_run as run
 import typer
 from typer.models import ParameterInfo
 
-from nemo_skills.pipeline.utils import get_mounts_from_config, get_tunnel
+from nemo_skills.pipeline.utils import (
+    get_mounts_from_config,
+    get_tunnel,
+    add_mount_path,
+    create_remote_directory,
+    check_remote_mount_directories,
+)  # noqa # pylint: disable=unused-import
 
 app = typer.Typer(no_args_is_help=True, pretty_exceptions_enable=False)
 
@@ -56,269 +62,3 @@ def typer_unpacker(f: Callable):
         return f(*args, **kwargs)
 
     return wrapper
-
-
-def create_remote_directory(directory: str | list, cluster_config: dict):
-    """Create a remote directory on the cluster."""
-
-    if cluster_config is None:
-        raise ValueError("Cluster config is not provided.")
-
-    if isinstance(directory, str):
-        directory = [directory]
-
-    if cluster_config.get('executor') != 'slurm':
-        tunnel = run.LocalTunnel(job_dir=directory[0])
-        for dir_path in directory:
-            tunnel.run(f'mkdir -p {dir_path}', hide=False, warn=True)
-            logging.info(f"Created directory: {dir_path} in local filesystem.")
-        tunnel.cleanup()
-
-    elif cluster_config.get('executor') == 'slurm':
-        ssh_tunnel_config = cluster_config.get('ssh_tunnel', None)
-        if ssh_tunnel_config is None:
-            raise ValueError("`ssh_tunnel` sub-config is not provided in cluster_config.")
-
-        # Check for pre-existing job_dir in the ssh_tunnel_config
-        if 'job_dir' not in ssh_tunnel_config:
-            ssh_tunnel_config['job_dir'] = directory[0]
-
-        tunnel = get_tunnel(cluster_config)
-        for dir_path in directory:
-            tunnel.run(f'mkdir -p {dir_path}', hide=False, warn=True)
-            logging.info(f"Created directory: {dir_path} on remote cluster.")
-        tunnel.cleanup()
-
-    else:
-        raise ValueError(f"Unsupported executor: {cluster_config.get('executor')}")
-
-
-def check_remote_mount_directories(directories: list, cluster_config: dict, exit_on_failure: bool = True):
-    """Create a remote directory on the cluster."""
-
-    if cluster_config is None:
-        raise ValueError("Cluster config is not provided.")
-
-    if isinstance(directories, str):
-        directories = [directories]
-
-    if cluster_config.get('executor') != 'slurm':
-        tunnel = run.LocalTunnel(job_dir=None)
-
-        all_dirs_exist = True
-        missing_source_locations = []
-        for directory in directories:
-            result = tunnel.run(f'test -e {directory} && echo "Directory Exists"', hide=True, warn=True)
-
-            if "Directory Exists" not in result.stdout:
-                missing_source_locations.append(directory)
-
-        tunnel.cleanup()
-
-        if len(missing_source_locations) > 0 and exit_on_failure:
-            missing_source_locations = [
-                f"{loc} DOES NOT exist at source destination" for loc in missing_source_locations
-            ]
-            missing_source_locations = "\n".join(missing_source_locations)
-            raise FileNotFoundError(
-                f"Some files or directories do not exist at the source location for mounting !!\n\n"
-                f"{missing_source_locations}"
-            )
-
-    elif cluster_config.get('executor') == 'slurm':
-        ssh_tunnel_config = cluster_config.get('ssh_tunnel', None)
-        if ssh_tunnel_config is None:
-            raise ValueError("`ssh_tunnel` sub-config is not provided in cluster_config.")
-
-        # Check for pre-existing job_dir in the ssh_tunnel_config
-        if 'job_dir' not in ssh_tunnel_config:
-            ssh_tunnel_config['job_dir'] = os.getcwd()
-
-        tunnel = get_tunnel(cluster_config)
-        missing_source_locations = []
-
-        for directory in directories:
-            result = tunnel.run(f'test -e {directory} && echo "Directory Exists"', hide=True, warn=True)
-
-            if "Directory Exists" not in result.stdout:
-                missing_source_locations.append(directory)
-
-        tunnel.cleanup()
-
-        if len(missing_source_locations) > 0 and exit_on_failure:
-            missing_source_locations = [
-                f"{loc} DOES NOT exist at source destination" for loc in missing_source_locations
-            ]
-            missing_source_locations = "\n".join(missing_source_locations)
-            raise FileNotFoundError(
-                f"Some files or directories do not exist at the source location for mounting !!\n\n"
-                f"{missing_source_locations}"
-            )
-
-    else:
-        raise ValueError(f"Unsupported executor: {cluster_config.get('executor')}")
-
-
-def add_mount_path(mount_source: str, mount_dest: str, cluster_config):
-    """Add a mount path to the cluster configuration."""
-
-    if cluster_config is None:
-        raise ValueError("Cluster config is not provided.")
-
-    if 'mounts' in cluster_config:
-        original_mounts = get_mounts_from_config(cluster_config)
-        added_mount = False
-        for mount_path in original_mounts:
-            source, destination = mount_path.split(':')
-
-            if source == mount_source and destination == mount_dest:
-                return
-
-        if not added_mount:
-            cluster_config['mounts'].append(f"{mount_source}:{mount_dest}")
-            logging.info(f"Added mount path: `{mount_source}:{mount_dest}`")
-
-    else:
-        raise ValueError("No mounts found in cluster config, can only add to existing mount list.")
-
-
-def is_mounted_filepath(filepath: str, cluster_config: dict):
-    """
-    Check if the filepath is mounted using the cluster config.
-
-    Args:
-        filepath: The filepath to check.
-        cluster_config: The cluster configuration dictionary.
-
-    Returns:
-        bool: Whether the filepath is mounted.
-    """
-    # Find which mount path matches the filepaths prefix
-    for mount in cluster_config['mounts']:
-        mount_source, mount_dest = mount.split(':')
-        if filepath.startswith(mount_dest):
-            return True
-
-    return False
-
-
-def get_mounted_filepath(filepath: str, cluster_config: dict):
-    """
-    Resolve the mounted filepath using the cluster config to merge the mount destination path to the filepath.
-    If an already mounted path is provided, then it will return the filepath as is.
-
-    Args:
-        filepath: The filepath to resolve.
-        cluster_config: The cluster configuration dictionary.
-
-    Returns:
-        The resolved filepath.
-    """
-    # Find which mount path matches the filepaths prefix
-    mount_path = None
-    for mount in cluster_config['mounts']:
-        mount_source, mount_dest = mount.split(':')
-        if filepath.startswith(mount_source):
-            mount_path = mount
-            break
-
-        elif filepath.startswith(mount_dest):
-            # already mounted, return immediately
-            return filepath
-
-    if mount_path is None:
-        raise ValueError(
-            f"Could not find a mount path for the file path `{filepath}`. Below paths are mounted: \n"
-            f"{cluster_config['mounts']}"
-        )
-
-    # replace the mount destination inside the filepath with the mount source
-    mount_source, mount_dest = mount_path.split(':')
-    filepath = mount_dest + filepath[len(mount_source) :]  # replace the mount destination with the mount source
-
-    return filepath
-
-
-def get_unmounted_filepath(filepath: str, cluster_config: dict,):
-    """
-    Resolve the mounted filepath using the cluster config to merge the mount source path to the filepath.
-    Args:
-        filepath: The filepath to resolve.
-        cluster_config: The cluster configuration dictionary.
-
-    Returns:
-        The resolved filepath.
-    """
-    # Find which mount path matches the filepaths prefix
-    mount_path = None
-    for mount in cluster_config['mounts']:
-        mount_source, mount_dest = mount.split(':')
-        if filepath.startswith(mount_dest):
-            mount_path = mount
-            break
-
-        elif filepath.startswith(mount_source):
-            # already mounted, return immediately
-            return filepath
-
-    if mount_path is None:
-        raise ValueError(
-            f"Could not find a mount path for the file path `{filepath}`. Below paths are mounted: \n"
-            f"{cluster_config['mounts']}"
-        )
-
-    # replace the mount destination inside the filepath with the mount source
-    mount_source, mount_dest = mount_path.split(':')
-    filepath = mount_source + filepath[len(mount_dest) :]  # replace the mount destination with the mount source
-
-    return filepath
-
-
-def resolve_mount_paths(mount_paths: str | list, cluster_config: dict, create_remote_dir: bool = True):
-    """
-    Resolve the mount paths using the cluster config to merge the mount destination path to the filepath.
-    Args:
-        mount_paths: The mount paths to resolve - can be a string (comma separated) or a list of strings.
-            Each mount path should be in the format `src:dest`.
-        cluster_config: The cluster configuration dictionary to update.
-        create_remote_dir: Whether to create the remote directories for the mount paths.
-
-    Return:
-        The updated cluster configuration dictionary.
-    """
-    logger = logging.getLogger('nemo_skills')
-
-    if mount_paths is not None:
-        if isinstance(mount_paths, str):
-            mount_paths_list = mount_paths.split(",")
-        else:
-            mount_paths_list = mount_paths
-        # remove empty strings from the list and strip whitespace
-        mount_paths_list = [path.strip() for path in mount_paths_list]
-        mount_paths_list = [path for path in mount_paths_list if path != ""]
-
-        for idx, path in enumerate(mount_paths_list):
-            assert ":" in path, f"Invalid mount path: {path}. Each path must be in the format `src:dest`"
-            src, dest = path.split(":")
-            src = src.strip().strip("\n")
-            dest = dest.strip().strip("\n")
-
-            logger.info(f"Dynamically mounting path: {src} to {dest}")
-            add_mount_path(src, dest, cluster_config)
-
-            # Incase `create_remote_dir` is set, we need to create the source directory.
-            # Preserve the original src path
-            mount_paths_list[idx] = (src, dest)
-
-        if create_remote_dir:
-            logger.info(f"Creating remote directories for mount paths:")
-            all_src_dir = [src for src, _ in mount_paths_list]
-            # Check if it is a file or a directory and only create the directory
-            for idx in range(len(all_src_dir)):
-                if os.path.splitext(all_src_dir[idx])[1] != "":
-                    all_src_dir[idx] = os.path.dirname(all_src_dir[idx])
-                logger.info(f"Creating remote directory: {all_src_dir[idx]}")
-
-            create_remote_directory(all_src_dir, cluster_config)
-
-    return cluster_config
